moved tests from separate dir to correspond packages
