--- test/calc_test/equal_float_test.go
@@ -1,9 +1,0 @@
-package calc_test
-
-import "math"
-
-const tolerance = 1e-8
-
-func equalFloat(num1, num2 float64) bool {
-	return math.Abs(num1-num2) < tolerance
-}
