--- test/uniq_test/nop_closer_buffer_test.go
@@ -1,11 +1,0 @@
-package uniq_test
-
-import "bytes"
-
-type nopCloseBuffer struct {
-	bytes.Buffer
-}
-
-func (nopCloseBuffer) Close() error {
-	return nil
-}
